Moves Admin Workstation pip reqs to central dir
Most of the project's various Python requirements files reside in
`securedrop/requirements/`, so let's store the Admin requirements there,
as well. Updated the securedrop-admin script to look in the new
location.

Removes unused `requirements.all` file for the Admin Workstation
dependencies, which was not referenced in any install logic or
documentation.

The generalized form of `ansible.in` is designed to allow a single
hardcoded version of Ansible to reused throughout multiple environments,
making testing more reliable.

Includes a slight alteration to the contents of the
admin-requirements.txt file, which is a no-op, since it's just the
comment provided by pip-compile during generation.