Consolidates develop-requirements files
The docs were needlessly recommending installation of two requirements
files during setup of the developer environment in a virtualenv. Since
we're using pip-compile, that's completely unnecessary. The CircleCI
logic similar duplicated requirements.

Now the contents of `testinfra/requirements.in` have been folded into
the `securedrop/requirements/develop-requirements.in` file, so only the
latter is required during setup (and CI).

Also referencing a single requirements file for the Ansible version,
so we don't have to update it in multiple locations all over the place.

Updated the Makefile target to generate these files in a reliable
fashion. Ran the new target in order to add pinned checksums for all of
the packages pulled in via the new consolidated develop-requirements
file.